refactor: Split aggregate.test.ts into separate files and extract shared helpers
