+++ .mulch/expertise/plots.jsonl
@@ -15,3 +15,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-23T16:49:56.448Z","evidence":{"seeds":"warren-3c12","commit":"dd134250ff46e88ba04ef9a060b96be96bfb7bcf"},"content":"defaultPlotSyncer (src/plots/sync.ts, warren-3c12 / pl-5a6c step 2) detects dirty .plot/ files via git status, creates a git worktree from origin/<targetBranch> to avoid modifying the main clone's branch, copies plot-*.json and plot-*.events.jsonl to it, commits as warren, pushes a unique warren/plot-sync-<hash> branch, opens a PR using openPullRequest, and optionally merges via mergePullRequest depending on the mergeStrategy (immediate/auto/manual).","id":"mx-bd189a"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-05-23T17:13:59.119Z","evidence":{"commit":"b6d49c7a74aa4181c77396b0e9b31d64f0344596"},"content":"plotsApi.sync(plotId) (src/ui/src/api/client.ts, warren-1d0c) is the typed fetch wrapper POSTing to /plots/:id/sync returning a PlotSyncResponse.","id":"mx-133277"}
 {"type":"pattern","classification":"tactical","recorded_at":"2026-05-25T19:30:42.073Z","evidence":{"seeds":"warren-d590","commit":"35604b50286865cdcbaf0aacaaf874df82a4bbe0"},"name":"queryWithRebuildRetry incremental stale-index detection","description":"queryWithRebuildRetry (src/plots/aggregate.ts) has three rebuild triggers: (1) error on first query → rebuild+retry, (2) zero rows + disk has *.json files → rebuild+retry (warren-ede7), (3) rows returned but disk has MORE *.json files than the index → rebuild+retry (warren-d590). The third trigger uses countPlotFilesOnDisk() on AggregatorPlotClient to compare index row count vs disk file count, catching incremental staleness when new plots land via git fetch after the index was built.","files":[".seeds/issues.jsonl","src/plots/aggregate.test.ts","src/plots/aggregate.ts"],"id":"mx-f9444e"}
+{"type":"convention","classification":"tactical","recorded_at":"2026-05-28T03:58:54.366Z","evidence":{"seeds":"warren-0d5a","commit":"88697321125d58a881d27b7d68673607ddf685e5"},"content":"Split src/plots/aggregate.test.ts (847 lines) into src/plots/aggregate.test.ts, src/plots/aggregate-needs-attention.test.ts, and src/plots/aggregate.test-helpers.ts to lower file size budgets","id":"mx-c0edf5"}
