+++ src/plots/aggregate.test.ts
@@ -16,174 +16,14 @@
  */
 
 import { describe, expect, test } from "bun:test";
-import type { PlotEvent, PlotStatus } from "@os-eco/plot-cli";
-import type { ProjectRow, RunRow } from "../db/schema.ts";
-import type { Logger } from "../server/types.ts";
 import {
-	type AggregatorClientFactory,
-	type AggregatorPlotClient,
-	type AggregatorRunsRepo,
-	createPlotAggregator,
-	EMPTY_PLOT_SUMMARIES,
-} from "./aggregate.ts";
-
-function silentLogger(): Logger {
-	return {
-		info() {},
-		warn() {},
-		error() {},
-	};
-}
-
-function captureLogger(): { logger: Logger; warns: Array<{ obj: object; msg?: string }> } {
-	const warns: Array<{ obj: object; msg?: string }> = [];
-	return {
-		warns,
-		logger: {
-			info() {},
-			warn(obj, msg) {
-				warns.push({ obj, msg });
-			},
-			error() {},
-		},
-	};
-}
-
-function project(id: string, hasPlot = true): ProjectRow {
-	return {
-		id,
-		gitUrl: `https://example.com/${id}.git`,
-		localPath: `/tmp/${id}`,
-		defaultBranch: "main",
-		addedAt: "2026-01-01T00:00:00Z",
-		lastFetchedAt: null,
-		lastHeadSha: null,
-		hasPlot,
-		hasSeeds: false,
-	};
-}
-
-interface StubPlot {
-	readonly id: string;
-	readonly name: string;
-	readonly status: PlotStatus;
-	readonly updated_at: string;
-	readonly goal: string;
-	readonly attachments: number;
-	readonly events: ReadonlyArray<PlotEvent>;
-}
-
-interface StubBehaviour {
-	readonly plots: ReadonlyArray<StubPlot>;
-	readonly failFirstQuery?: boolean;
-	readonly failAllQueries?: boolean;
-	/**
-	 * Plots that exist on disk as `*.json` files but are missing from
-	 * the index DB. When set, `query()` returns rows derived from
-	 * `plots` (typically empty until `rebuildIndex` runs), and
-	 * `rebuildIndex` migrates these into `plots` so the retry sees
-	 * them. Models the warren-ede7 cold-cache empty-rows path.
-	 */
-	readonly plotsOnDiskOnly?: ReadonlyArray<StubPlot>;
-	/**
-	 * Override the disk probe result. When `plotsOnDiskOnly` is set
-	 * this defaults to `true`; otherwise to `plots.length > 0`.
-	 */
-	readonly hasFilesOnDisk?: boolean;
-}
-
-interface StubMetrics {
-	queryCalls: number;
-	rebuildCalls: number;
-	closeCalls: number;
-	hasFilesOnDiskCalls: number;
-	countFilesOnDiskCalls: number;
-}
-
-function makeFactory(perProject: Record<string, StubBehaviour>): {
-	factory: AggregatorClientFactory;
-	metrics: Record<string, StubMetrics>;
-} {
-	const metrics: Record<string, StubMetrics> = {};
-	const factory: AggregatorClientFactory = (p) => {
-		const behaviour = perProject[p.id];
-		if (behaviour === undefined) {
-			throw new Error(`unexpected project in factory: ${p.id}`);
-		}
-		let m = metrics[p.id];
-		if (m === undefined) {
-			m = {
-				queryCalls: 0,
-				rebuildCalls: 0,
-				closeCalls: 0,
-				hasFilesOnDiskCalls: 0,
-				countFilesOnDiskCalls: 0,
-			};
-			metrics[p.id] = m;
-		}
-		// Live view of indexed plots — rebuildIndex absorbs the
-		// `plotsOnDiskOnly` set so the retry query sees them.
-		const indexed: StubPlot[] = [...behaviour.plots];
-		const onDiskOnly: StubPlot[] = behaviour.plotsOnDiskOnly ? [...behaviour.plotsOnDiskOnly] : [];
-		const client: AggregatorPlotClient = {
-			async query() {
-				m.queryCalls += 1;
-				if (behaviour.failAllQueries) throw new Error("query-broken");
-				if (behaviour.failFirstQuery && m.queryCalls === 1) {
-					throw new Error("first-query-broken");
-				}
-				return { rows: indexed.map((pl) => ({ id: pl.id })) };
-			},
-			async rebuildIndex() {
-				m.rebuildCalls += 1;
-				while (onDiskOnly.length > 0) {
-					const next = onDiskOnly.shift();
-					if (next !== undefined) indexed.push(next);
-				}
-			},
-			async hasPlotFilesOnDisk() {
-				m.hasFilesOnDiskCalls += 1;
-				if (behaviour.hasFilesOnDisk !== undefined) return behaviour.hasFilesOnDisk;
-				if (onDiskOnly.length > 0 || indexed.length > 0) return true;
-				return false;
-			},
-			async countPlotFilesOnDisk() {
-				m.countFilesOnDiskCalls += 1;
-				return indexed.length + onDiskOnly.length;
-			},
-			async readPlot(plotId) {
-				const pl = indexed.find((x) => x.id === plotId);
-				if (pl === undefined) throw new Error(`unknown plot ${plotId}`);
-				return {
-					name: pl.name,
-					status: pl.status,
-					updated_at: pl.updated_at,
-					intent: { goal: pl.goal },
-					attachments: new Array(pl.attachments).fill(null),
-				};
-			},
-			async readEvents(plotId) {
-				const pl = indexed.find((x) => x.id === plotId);
-				if (pl === undefined) throw new Error(`unknown plot ${plotId}`);
-				return pl.events;
-			},
-			close() {
-				m.closeCalls += 1;
-			},
-		};
-		return client;
-	};
-	return { factory, metrics };
-}
-
-function noteEvent(at: string, actor: string): PlotEvent {
-	return {
-		type: "note",
-		actor,
-		at,
-		data: { text: "x" },
-	};
-}
+	captureLogger,
+	makeFactory,
+	noteEvent,
+	project,
+	silentLogger,
+} from "./aggregate.test-helpers.ts";
+import { createPlotAggregator, EMPTY_PLOT_SUMMARIES } from "./aggregate.ts";
 
 describe("createPlotAggregator", () => {
 	test("returns the canonical EMPTY reference when no project has hasPlot=true", async () => {
@@ -581,267 +421,5 @@ describe("createPlotAggregator", () => {
 		expect(rows[0]?.last_event_ts).toBe("2026-05-01T12:00:00Z");
 		expect(rows[0]?.last_event_actor).toBe("");
 		expect(rows[0]?.intent_goal_preview).toBe("");
-	});
-});
-
-/**
- * Needs-attention scorer integration (warren-d693 / pl-0344 step 9).
- * The pure per-Plot policy lives in `./needs-attention.ts` and has its
- * own unit tests there; here we pin the aggregator-level wiring: paused
- * runs grouped from `listByState('paused')`, stale-draft window applied
- * with the injected clock, and the count endpoint mirroring the list.
- */
-function pausedRunsRepo(plotIds: readonly (string | null)[]): AggregatorRunsRepo {
-	return {
-		async listByState(state) {
-			expect(state).toBe("paused");
-			return plotIds.map((id) => ({ plotId: id }) as unknown as RunRow);
-		},
-	};
-}
-
-describe("createPlotAggregator listNeedsAttention", () => {
-	test("flags Plots with a paused run on the runsRepo", async () => {
-		const projects = [project("prj_a")];
-		const { factory } = makeFactory({
-			prj_a: {
-				plots: [
-					{
-						id: "plot-paused1",
-						name: "p",
-						status: "active",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-05-20T00:00:00Z", "user:a")],
-					},
-					{
-						id: "plot-quiet1",
-						name: "q",
-						status: "active",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-05-20T00:00:00Z", "user:a")],
-					},
-				],
-			},
-		});
-		const agg = createPlotAggregator({
-			projectsRepo: { listAll: async () => projects },
-			logger: silentLogger(),
-			clientFactory: factory,
-			runsRepo: pausedRunsRepo(["plot-paused1", null]),
-			now: () => Date.parse("2026-05-21T00:00:00Z"),
-		});
-		const rows = await agg.listNeedsAttention();
-		expect(rows.map((r) => r.id)).toEqual(["plot-paused1"]);
-		expect(rows[0]?.reasons).toEqual(["paused_run"]);
-		expect(await agg.countNeedsAttention()).toBe(1);
-	});
-
-	test("flags drafting Plots whose last_event_ts is older than the stale window", async () => {
-		const projects = [project("prj_a")];
-		const { factory } = makeFactory({
-			prj_a: {
-				plots: [
-					{
-						id: "plot-stale1",
-						name: "old draft",
-						status: "drafting",
-						updated_at: "2026-05-01T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-05-01T00:00:00Z", "user:a")],
-					},
-					{
-						id: "plot-fresh1",
-						name: "fresh draft",
-						status: "drafting",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-05-20T00:00:00Z", "user:a")],
-					},
-					{
-						id: "plot-doneish",
-						name: "non-draft",
-						status: "active",
-						updated_at: "2026-04-01T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-04-01T00:00:00Z", "user:a")],
-					},
-				],
-			},
-		});
-		const agg = createPlotAggregator({
-			projectsRepo: { listAll: async () => projects },
-			logger: silentLogger(),
-			clientFactory: factory,
-			runsRepo: pausedRunsRepo([]),
-			staleDraftAfterDays: 7,
-			// 21 days after plot-stale1's last event, 1 day after plot-fresh1's.
-			now: () => Date.parse("2026-05-21T00:00:00Z"),
-		});
-		const rows = await agg.listNeedsAttention();
-		expect(rows.map((r) => r.id)).toEqual(["plot-stale1"]);
-		expect(rows[0]?.reasons).toEqual(["stale_draft"]);
-	});
-
-	test("flags merged gh_pr attachments with no follow-up review event", async () => {
-		const projects = [project("prj_a")];
-		const { factory } = makeFactory({
-			prj_a: {
-				plots: [
-					{
-						id: "plot-pr1",
-						name: "merged pr unreviewed",
-						status: "active",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 1,
-						events: [
-							{
-								type: "artifact_produced",
-								actor: "agent:claude-code:run_x",
-								at: "2026-05-20T00:00:00Z",
-								data: { type: "gh_pr", ref: "owner/repo#42" },
-							} as PlotEvent,
-						],
-					},
-					{
-						id: "plot-pr2",
-						name: "merged pr reviewed",
-						status: "active",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 1,
-						events: [
-							{
-								type: "artifact_produced",
-								actor: "agent:x",
-								at: "2026-05-20T00:00:00Z",
-								data: { type: "gh_pr", ref: "owner/repo#43" },
-							} as PlotEvent,
-							{
-								type: "decision_made",
-								actor: "user:operator",
-								at: "2026-05-21T00:00:00Z",
-								data: { summary: "reviewed owner/repo#43, looks good" },
-							} as PlotEvent,
-						],
-					},
-				],
-			},
-		});
-		const agg = createPlotAggregator({
-			projectsRepo: { listAll: async () => projects },
-			logger: silentLogger(),
-			clientFactory: factory,
-			runsRepo: pausedRunsRepo([]),
-			now: () => Date.parse("2026-05-22T00:00:00Z"),
-		});
-		const rows = await agg.listNeedsAttention();
-		expect(rows.map((r) => r.id)).toEqual(["plot-pr1"]);
-		expect(rows[0]?.reasons).toEqual(["merged_pr_unreviewed"]);
-	});
-
-	test("returns multiple reasons in canonical order for a single Plot", async () => {
-		const projects = [project("prj_a")];
-		const { factory } = makeFactory({
-			prj_a: {
-				plots: [
-					{
-						id: "plot-multi",
-						name: "multi-signal",
-						status: "drafting",
-						updated_at: "2026-05-01T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [
-							noteEvent("2026-05-01T00:00:00Z", "user:a"),
-							{
-								type: "artifact_produced",
-								actor: "agent:x",
-								at: "2026-05-01T01:00:00Z",
-								data: { type: "gh_pr", ref: "owner/repo#1" },
-							} as PlotEvent,
-						],
-					},
-				],
-			},
-		});
-		const agg = createPlotAggregator({
-			projectsRepo: { listAll: async () => projects },
-			logger: silentLogger(),
-			clientFactory: factory,
-			runsRepo: pausedRunsRepo(["plot-multi"]),
-			staleDraftAfterDays: 7,
-			now: () => Date.parse("2026-05-21T00:00:00Z"),
-		});
-		const rows = await agg.listNeedsAttention();
-		expect(rows[0]?.reasons).toEqual(["paused_run", "merged_pr_unreviewed", "stale_draft"]);
-	});
-
-	test("returns [] when no Plot qualifies and countNeedsAttention agrees", async () => {
-		const projects = [project("prj_a")];
-		const { factory } = makeFactory({
-			prj_a: {
-				plots: [
-					{
-						id: "plot-ok",
-						name: "ok",
-						status: "active",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-05-20T00:00:00Z", "user:a")],
-					},
-				],
-			},
-		});
-		const agg = createPlotAggregator({
-			projectsRepo: { listAll: async () => projects },
-			logger: silentLogger(),
-			clientFactory: factory,
-			runsRepo: pausedRunsRepo([]),
-			now: () => Date.parse("2026-05-21T00:00:00Z"),
-		});
-		expect(await agg.listNeedsAttention()).toEqual([]);
-		expect(await agg.countNeedsAttention()).toBe(0);
-	});
-
-	test("tolerates a runsRepo query failure (logs + treats as zero paused runs)", async () => {
-		const projects = [project("prj_a")];
-		const { factory } = makeFactory({
-			prj_a: {
-				plots: [
-					{
-						id: "plot-x",
-						name: "x",
-						status: "active",
-						updated_at: "2026-05-20T00:00:00Z",
-						goal: "g",
-						attachments: 0,
-						events: [noteEvent("2026-05-20T00:00:00Z", "user:a")],
-					},
-				],
-			},
-		});
-		const cap = captureLogger();
-		const agg = createPlotAggregator({
-			projectsRepo: { listAll: async () => projects },
-			logger: cap.logger,
-			clientFactory: factory,
-			runsRepo: {
-				async listByState() {
-					throw new Error("runs db down");
-				},
-			},
-			now: () => Date.parse("2026-05-21T00:00:00Z"),
-		});
-		expect(await agg.listNeedsAttention()).toEqual([]);
-		expect(cap.warns.some((w) => w.msg === "plots.needs_attention_paused_query_failed")).toBe(true);
 	});
 });
